Multi-file upload capability added.

+++ s3backup-cli/trunk/src/main/java/org/wiztools/site/s3backup/Main.java
@@ -16,6 +16,7 @@ import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 import org.jets3t.service.S3Service;
 import org.jets3t.service.S3ServiceException;
+import org.jets3t.service.ServiceException;
 import org.jets3t.service.impl.rest.httpclient.RestS3Service;
 import org.jets3t.service.model.S3Bucket;
 import org.jets3t.service.model.S3Object;
@@ -27,6 +28,11 @@ import org.jets3t.service.security.AWSCredentials;
  */
 public class Main {
     
+    private static final int EXIT_CLI_ERROR = 1;
+    private static final int EXIT_IO_ERROR = 2;
+    private static final int EXIT_S3_ERROR = 3;
+    private static final int EXIT_SYS_ERROR = 4;
+    
     private static Options generateOptions(){
         Options options = new Options();
         
@@ -34,6 +40,7 @@ public class Main {
         Option option = OptionBuilder
                 .withLongOpt("aws-creds-file")
                 .hasArg()
+                .withDescription("Java properties file with AWS credentials")
                 .create('k');
         options.addOption(option);
         
@@ -41,6 +48,7 @@ public class Main {
         option = OptionBuilder
                 .withLongOpt("accesskey")
                 .hasArg()
+                .withDescription("AWS access key (not needed when -k option is used)")
                 .create('a');
         options.addOption(option);
         
@@ -48,6 +56,7 @@ public class Main {
         option = OptionBuilder
                 .withLongOpt("secretkey")
                 .hasArg()
+                .withDescription("AWS secret key (not needed when -k option is used)")
                 .create('s');
         options.addOption(option);
         
@@ -56,15 +65,8 @@ public class Main {
                 .withLongOpt("bucket")
                 .hasArg()
                 .isRequired()
+                .withDescription("Destination S3 bucket name")
                 .create('b');
-        options.addOption(option);
-        
-        // filename
-        option = OptionBuilder
-                .withLongOpt("file")
-                .hasArg()
-                .isRequired()
-                .create('f');
         options.addOption(option);
         
         // help
@@ -72,6 +74,7 @@ public class Main {
                 .withLongOpt("help")
                 .hasArg(false)
                 .isRequired(false)
+                .withDescription("Print this help")
                 .create('h');
         options.addOption(option);
         
@@ -80,7 +83,7 @@ public class Main {
     
     private static void printCommandLineHelp(Options options){
         HelpFormatter hf = new HelpFormatter();
-        String cmdLine = "java -jar s3backup-NN-jar-with-dependencies.jar [options]";
+        String cmdLine = "java -jar s3backup-NN-jar-with-dependencies.jar [options] file1 file2 ...";
         String descriptor = "AWS S3 backup tool";
         String moreHelp = "Format of `aws-creds-file': \n"
                 + "\tAWSAccessKeyId=XXX\n"
@@ -98,23 +101,28 @@ public class Main {
                 printCommandLineHelp(options);
                 return;
             }
+            
+            if(cmd.getArgs().length == 0) {
+                System.err.println("No files specified for transfer.");
+                printCommandLineHelp(options);
+                System.exit(EXIT_CLI_ERROR);
+            }
 
             String awsCredsFile = options.getOption("aws-creds-file").getValue();
             String accessKey = options.getOption("accesskey").getValue();
             String secretKey = options.getOption("secretkey").getValue();
             String bucketName = options.getOption("bucket").getValue();
-            String fileName = options.getOption("file").getValue();
             
             if(awsCredsFile == null && (accessKey == null || secretKey == null)) {
                 System.err.println("Either -k or (-a and -s) options are mandatory.");
                 printCommandLineHelp(options);
-                System.exit(1);
+                System.exit(EXIT_CLI_ERROR);
             }
             
             if(awsCredsFile != null && (accessKey != null && secretKey != null)) {
                 System.err.println("Options -k and (-a and -s) cannot coexist.");
                 printCommandLineHelp(options);
-                System.exit(2);
+                System.exit(EXIT_CLI_ERROR);
             }
             
             if(awsCredsFile != null) {
@@ -128,7 +136,7 @@ public class Main {
                 catch(IOException ex) {
                     System.err.println("Cannot read AWS property file.");
                     ex.printStackTrace(System.err);
-                    System.exit(3);
+                    System.exit(EXIT_IO_ERROR);
                 }
             }
             
@@ -136,30 +144,46 @@ public class Main {
             try{
                 S3Service service = new RestS3Service(cred);
                 S3Bucket bucket = new S3Bucket(bucketName);
-                File file = new File(fileName);
-                S3Object object = new S3Object(bucket, file);
-                service.putObject(bucket, object);
+                
+                for(String fileName: cmd.getArgs()) {
+                    File file = new File(fileName);
+                    if(file.isFile() && file.canRead()) {
+                        S3Object object = new S3Object(bucket, file);
+                        service.putObject(bucket, object);
+                    }
+                    else {
+                        System.err.println("Failed: " + fileName);
+                    }
+                }
+                
+                service.shutdown();
+                
             }
             catch(S3ServiceException ex){
                 ex.printStackTrace(System.err);
-                System.exit(4);
+                System.exit(EXIT_S3_ERROR);
+            }
+            catch(ServiceException ex) {
+                ex.printStackTrace(System.err);
+                System.exit(EXIT_S3_ERROR);
             }
             catch(NoSuchAlgorithmException ex){
                 ex.printStackTrace(System.err);
-                System.exit(5);
+                System.exit(EXIT_SYS_ERROR);
             }
             catch(FileNotFoundException ex){
                 ex.printStackTrace(System.err);
-                System.exit(6);
+                System.exit(EXIT_IO_ERROR);
             }
             catch(IOException ex){
                 ex.printStackTrace(System.err);
-                System.exit(7);
+                System.exit(EXIT_IO_ERROR);
             }
         }
         catch(ParseException ex){
+            System.err.println(ex.getMessage());
             printCommandLineHelp(options);
-            System.exit(8);
+            System.exit(EXIT_CLI_ERROR);
         }
     }
 }
